search: Move FakeClock to timeutil
This moves `FakeClock` from `dbtesting` to `timeutil` so it can used
without introducing a transitive dependency on `dbconn`.
