+++ cmd/repo-updater/repoupdater/server_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	idb "github.com/sourcegraph/sourcegraph/internal/db"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbtest"
-	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/awscodecommit"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketserver"
@@ -33,6 +32,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/repos"
 	"github.com/sourcegraph/sourcegraph/internal/repoupdater"
 	"github.com/sourcegraph/sourcegraph/internal/repoupdater/protocol"
+	"github.com/sourcegraph/sourcegraph/internal/timeutil"
 	"github.com/sourcegraph/sourcegraph/internal/trace"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 )
@@ -770,7 +770,7 @@ func testServerStatusMessages(t *testing.T, store *repos.Store) func(t *testing.
 					t.Fatal(err)
 				}
 
-				clock := dbtesting.NewFakeClock(time.Now(), 0)
+				clock := timeutil.NewFakeClock(time.Now(), 0)
 				syncer := &repos.Syncer{
 					Store: store,
 					Now:   clock.Now,
@@ -846,7 +846,7 @@ func testRepoLookup(db *sql.DB) func(t *testing.T, repoStore *repos.Store) func(
 	return func(t *testing.T, store *repos.Store) func(t *testing.T) {
 		return func(t *testing.T) {
 			ctx := context.Background()
-			clock := dbtesting.NewFakeClock(time.Now(), 0)
+			clock := timeutil.NewFakeClock(time.Now(), 0)
 			now := clock.Now()
 
 			githubSource := types.ExternalService{
