+++ internal/repos/syncer_test.go
@@ -18,7 +18,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	idb "github.com/sourcegraph/sourcegraph/internal/db"
-	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/awscodecommit"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketcloud"
@@ -27,6 +26,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/gitlab"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/gitolite"
 	"github.com/sourcegraph/sourcegraph/internal/repos"
+	"github.com/sourcegraph/sourcegraph/internal/timeutil"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 	"github.com/sourcegraph/sourcegraph/schema"
 )
@@ -87,7 +87,7 @@ func testSyncerSyncWithErrors(t *testing.T, store *repos.Store) func(t *testing.
 		} {
 			tc := tc
 			t.Run(tc.name, func(t *testing.T) {
-				clock := dbtesting.NewFakeClock(time.Now(), time.Second)
+				clock := timeutil.NewFakeClock(time.Now(), time.Second)
 				now := clock.Now
 				ctx := context.Background()
 
@@ -230,7 +230,7 @@ func testSyncerSync(t *testing.T, s *repos.Store) func(*testing.T) {
 		types.Opt.RepoSources(bitbucketCloudService.URN()),
 	)
 
-	clock := dbtesting.NewFakeClock(time.Now(), 0)
+	clock := timeutil.NewFakeClock(time.Now(), 0)
 
 	svcdup := types.ExternalService{
 		Kind:        extsvc.KindGitHub,
@@ -624,7 +624,7 @@ func testSyncerSync(t *testing.T, s *repos.Store) func(*testing.T) {
 
 				now := tc.now
 				if now == nil {
-					clock := dbtesting.NewFakeClock(time.Now(), time.Second)
+					clock := timeutil.NewFakeClock(time.Now(), time.Second)
 					now = clock.Now
 				}
 
@@ -676,7 +676,7 @@ func testSyncerSync(t *testing.T, s *repos.Store) func(*testing.T) {
 }
 
 func testSyncRepo(t *testing.T, s *repos.Store) func(*testing.T) {
-	clock := dbtesting.NewFakeClock(time.Now(), time.Second)
+	clock := timeutil.NewFakeClock(time.Now(), time.Second)
 
 	servicesPerKind := createExternalServices(t, s)
